requirements.txt
django==6.0
djangorestframework==3.16.1
django-environ==0.12.0
psycopg[c]==3.3.2
gunicorn==23.0.0
channels[daphne]==4.3.2
channels-redis==4.3.0
daphne==4.2.1
argon2-cffi==25.1.0
whitenoise==6.11.0
django-storages[boto3]==1.14.6
celery==5.6.0
redis==7.1.0
openai==2.11.0
pydantic==2.12.5
httpx==0.28.1
